virttest.utils_test: Add libvirt submodule
By putting functions more focused in the libvirt tests,
we have a better organized API.
