+++ libvirt/tests/src/virsh_cmd/domain/virsh_emulatorpin.py
@@ -4,7 +4,7 @@ import logging
 from virttest.libvirt_xml import vm_xml
 from autotest.client.shared import error
 from virttest import utils_libvirtd, virsh
-from virttest.utils_test import cpus_parser
+from virttest.utils_test.libvirt import cpus_parser
 
 try:
     from virttest.staging import utils_cgroup
